Move `FeedImageDataLoadTask` protocol and `LocalFeedImageDataLoader` to production
