+++ FeedFetcher/FeedFetcherTests/FeedCache/LoadFeedImageFromCacheUseCaseTests.swift
@@ -5,74 +5,6 @@ import Foundation
 import XCTest
 import FeedFetcher
 
-protocol FeedImageDataLoadTask {
-    func cancel()
-}
-
-final class LocalFeedImageDataLoader {
-    typealias Result = Swift.Result<Data, Error>
-    enum Error: Swift.Error {
-        case loadingImageData
-        case notFound
-    }
-    
-    private let store: FeedImageDataStore
-    
-    init(store: FeedImageDataStore) {
-        self.store = store
-    }
-    
-    @discardableResult
-    func loadImageData(for url: URL, completion: @escaping (Result) -> Void) -> FeedImageDataLoadTask {
-        let task = ImageDataLoadTask(completion: completion)
-        
-        task.wrappedTask =  store.retrieveImageData(for: url) { [weak self] retrieveResult in
-            guard let self = self else { return }
-            
-            let loadResult = self.handle(storeRetrieveResult: retrieveResult)
-            task.complete(with: loadResult)
-        }
-        
-        return task
-    }
-    
-    private func handle(storeRetrieveResult retrieveResult: FeedImageDataStore.Result) -> LocalFeedImageDataLoader.Result {
-        let loadResult = retrieveResult
-            .mapError{ _ in Error.loadingImageData}
-            .flatMap{ data in
-                return !data.isEmpty ? .success(data) :.failure(.notFound)
-            }
-        
-        return loadResult
-    }
-    
-    // MARK:- ImageDataLoadTask
-    
-    private class ImageDataLoadTask: FeedImageDataLoadTask {
-        private var completion: ((LocalFeedImageDataLoader.Result) -> Void)?
-        var wrappedTask: RetrieveImageDataTask?
-        
-        init(completion: @escaping (LocalFeedImageDataLoader.Result) -> Void) {
-            self.completion = completion
-        }
-        
-        func cancel() {
-            wrappedTask?.cancel()
-            invalidateCompletion()
-        }
-        
-        func complete(with result: LocalFeedImageDataLoader.Result) {
-            completion?(result)
-        }
-        
-        private func invalidateCompletion() {
-            completion = nil
-        }
-    }
-}
-
-
-
 class LoadFeedImageDataFromCacheUseCaseTests: XCTestCase {
     
     func test_feedImageDataStore_doesNotRequestDataAtInit() {
